audit-ai-system/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
langchain
langchain-community
langgraph
pypdf
tiktoken
pgvector
psycopg2-binary
pydantic
python-multipart
